Fixed SlashCommandHandler Location
For some reason, i was editing a ghost file that wasn't in the right place. Unsure how but this should rectify the source.
