+++ dGame/dUtilities/SlashCommandHandler.cpp
@@ -62,6 +62,22 @@
 #include "VanityUtilities.h"
 #include "GameConfig.h"
 
+std::string gen_random(const int len) {
+    static const char alphanum[] =
+        "0123456789"
+        "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
+        "abcdefghijklmnopqrstuvwxyz";
+    std::string tmp_s;
+    tmp_s.reserve(len);
+
+    for (int i = 0; i < len; ++i) {
+        tmp_s += alphanum[rand() % (sizeof(alphanum) - 1)];
+    }
+
+    return tmp_s;
+}
+
+
 void SlashCommandHandler::HandleChatCommand(const std::u16string& command, Entity* entity, const SystemAddress& sysAddr) {
     std::string chatCommand;
     std::vector<std::string> args;
@@ -162,6 +178,59 @@ void SlashCommandHandler::HandleChatCommand(const std::u16string& command, Entit
 	//!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 	//HANDLE ALL NON GM SLASH COMMANDS RIGHT HERE!
 	//!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+	 
+	if (chatCommand == "verify") { // Redi's Custom Command for Verification Interactions
+            Game::logger->Log("SlashCommandHandler", "I was called!");
+            std::stringstream test;
+            test << "Your latest message has been heard!";
+
+            ChatPackets::SendSystemMessage(sysAddr, GeneralUtils::ASCIIToUTF16(test.str()));
+
+			uint32_t id = entity->GetCharacter()->GetID();
+            std::string code = "C_" + gen_random(10);
+			
+			// Check if the verification stage is already done or not.
+			// Displays the Discord ID fetched. 0 Means default/not verified.
+            std::string discordid; 
+            auto grabstmt = Database::CreatePreppedStmt("SELECT verify_id FROM discord_verification WHERE acc_id = ? LIMIT 1;");
+            grabstmt->setInt(1, id);
+
+            sql::ResultSet* res = grabstmt->executeQuery();
+            
+			if (res->rowsCount() != 1) { // This means that the verification records do not exist. (Hasn't ever been executed.)
+                discordid = "0";
+                auto insertdb = Database::CreatePreppedStmt("INSERT INTO discord_verification (acc_id, verify_id, verify_code) VALUES (?, ?, ?)");
+                insertdb->setInt(1, id);
+                insertdb->setString(2, "0");
+                insertdb->setString(3, code);
+                insertdb->execute();
+				delete insertdb;
+            } else if (res->next()) {
+                discordid = res->getString(1);
+            }
+
+			delete res;
+
+			if (discordid != "0") {
+                ChatPackets::SendSystemMessage(sysAddr, u"Your account has already been verified. If you have any questions, please raise it with a Mythran @ Luplo Discord.");
+				return;
+			}
+
+			delete grabstmt;
+
+			// Overwrite code with a new one.
+			auto updatestmt = Database::CreatePreppedStmt("UPDATE discord_verification SET verify_code = ? WHERE acc_id = ?;");
+            updatestmt->setString(1, code);
+            updatestmt->setInt(2, id);
+            updatestmt->execute();
+            delete updatestmt;
+
+            std::stringstream message;
+            message << "Your verification code is [" << code << "]. Verify by sending the code to the bot's DMs.";
+
+			ChatPackets::SendSystemMessage(sysAddr, GeneralUtils::ASCIIToUTF16(message.str()));
+            return;
+	}
 
 	if (chatCommand == "pvp") {
 		auto* character = entity->GetComponent<CharacterComponent>();
@@ -1961,4 +2030,3 @@ void SlashCommandHandler::SendAnnouncement(const std::string& title, const std::
 
 	Game::chatServer->Send(&bitStream, SYSTEM_PRIORITY, RELIABLE, 0, Game::chatSysAddr, false);
 }
-
